allowing publishing empty point clouds
unfortunately, I can't remember why I introduced in the first
place this check. I remember something was crashing. However,
things are working now, and it makes sense to publish an
empty point cloud in the case the result of the query is... empty.
You might see me reverting this commit in the future.